refactored to use twa components + removed managing button state
