+++ src/store.ts
@@ -242,53 +242,12 @@ export const useErrorStore = create<ErrorStore>((set, get) => ({
 }));
 
 
-export enum TwaButtonType {
-  Connect = "Connect",
-  CastVote = "CastVote",
-  ConfirmVote = "ConfirmVote",
-}
-
-type SetMainButtonProps = {
-  twaButtonType: TwaButtonType
-  clickHandler: () => void
-  text: string
-}
-
-type HideMainButtonProps = {
-  clickHandler: () => void
-}
-
 interface TwaStore {
   isTwa?: boolean;
   setIsTwa: (isTwa: boolean | undefined) => void;
-  twaButtonType?: TwaButtonType;
-  setMainButton: (props: SetMainButtonProps) => void;
-  hideMainButton: (props: HideMainButtonProps) => void;
 }
 
 export const useTwaStore = create<TwaStore>((set, get) => ({
   isTwa: undefined,
   setIsTwa: (isTwa) => set({ isTwa }),
-  twaButtonType: undefined,
-  setMainButton: ({twaButtonType, clickHandler, text}) => {
-
-    const currentButtonType = get().twaButtonType
-
-    if (currentButtonType === TwaButtonType.Connect) {
-      return
-    }
-
-    twa.MainButton.onClick(clickHandler)
-    twa.MainButton.setParams({
-      text,
-      is_visible: true,
-    })
-    set({ twaButtonType })
-
-  },
-  hideMainButton: ({clickHandler}) => {
-    twa.MainButton.offClick(clickHandler)
-    twa.MainButton.hide()
-    set({ twaButtonType: undefined })
-  }
 }));
